+++ studio-platform-user-default/src/main/java/studio/one/base/user/web/controller/UserMgmtController.java
@@ -76,7 +76,7 @@ import studio.one.platform.web.dto.ApiResponse;
 @RequestMapping("${" + PropertyKeys.Features.User.Web.BASE_PATH + ":/api/mgmt}/users")
 @RequiredArgsConstructor
 @Slf4j
-public class UserMgmtController extends AbstractPasswordPolicyControllerSupport implements UserMgmtControllerApi {
+public class UserMgmtController extends AbstractPasswordPolicyControllerSupport implements UserMgmtApi {
 
         private final ApplicationUserService<User, Role> userService;
         private final ApplicationUserMapper userMapper;